Make conversion behavoir depend on shell type.
The shell type is responsible for what conversions need to be
applied to the command line. MSYS bash and cygwin bash have different
behaviour when it comes to interpreting the commandline
switches. Therefore omnisharp--shell-type test has been introduced a
variable that caches the test result and the conversion has been made
sensitive to the shell-type.